+++ Code/Tardigrade.Framework.Tests/Tardigrade.Framework.Tests.sln
@@ -41,7 +41,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Tardigrade.Framework.MailKi
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Tardigrade.Framework.EntityFramework.Tests", "Tardigrade.Framework.EntityFramework.Tests\Tardigrade.Framework.EntityFramework.Tests.csproj", "{64B8F9B6-FA73-4566-8CF6-006303761626}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Tardigrade.Framework.AuditNET.Tests", "Tardigrade.Framework.AuditNET.Tests\Tardigrade.Framework.AuditNET.Tests.csproj", "{5B0B14DF-B012-4CBB-AC7F-8226A25AC459}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Tardigrade.Framework.AuditNET.Tests", "Tardigrade.Framework.AuditNET.Tests\Tardigrade.Framework.AuditNET.Tests.csproj", "{5B0B14DF-B012-4CBB-AC7F-8226A25AC459}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Tardigrade.Framework.AuditNET.AzureStorageQueue", "..\Tardigrade.Framework\Tardigrade.Framework.AuditNET.AzureStorageQueue\Tardigrade.Framework.AuditNET.AzureStorageQueue.csproj", "{5B90F40B-74CC-40F1-A04D-E952CFF7CF3C}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -129,6 +131,10 @@ Global
 		{5B0B14DF-B012-4CBB-AC7F-8226A25AC459}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5B0B14DF-B012-4CBB-AC7F-8226A25AC459}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{5B0B14DF-B012-4CBB-AC7F-8226A25AC459}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5B90F40B-74CC-40F1-A04D-E952CFF7CF3C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5B90F40B-74CC-40F1-A04D-E952CFF7CF3C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5B90F40B-74CC-40F1-A04D-E952CFF7CF3C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5B90F40B-74CC-40F1-A04D-E952CFF7CF3C}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
