+++ Code/Tardigrade.Framework/Tardigrade.Framework.AuditNET.AzureStorageQueue/Tardigrade.Framework.AuditNET.AzureStorageQueue.csproj
@@ -2,9 +2,9 @@
 
   <PropertyGroup>
     <TargetFrameworks>net6.0;netstandard2.0</TargetFrameworks>
-    <Version>6.0.0</Version>
+    <Version>1.0.0</Version>
     <Authors>Rafidzal Rafiq</Authors>
-    <Product>Tardigrade Framework Library (Audit.NET Integration)</Product>
+    <Product>Tardigrade Framework Library (Audit.NET/Azure Storage Queue Integration)</Product>
     <Description>Software framework for supporting coding best practices.</Description>
     <Copyright>Copyright © 2022</Copyright>
     <PackageLicenseExpression>Apache-2.0</PackageLicenseExpression>
@@ -16,11 +16,11 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Audit.NET" Version="19.0.5" />
+    <PackageReference Include="Azure.Storage.Queues" Version="12.8.0" />
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Tardigrade.Framework\Tardigrade.Framework.csproj" />
+    <ProjectReference Include="..\Tardigrade.Framework.AuditNET\Tardigrade.Framework.AuditNET.csproj" />
   </ItemGroup>
 
 </Project>
