+++ Code/Tardigrade.Framework/Tardigrade.Framework.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.29001.49
+# Visual Studio Version 17
+VisualStudioVersion = 17.1.32210.238
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Tardigrade.Framework", "Tardigrade.Framework\Tardigrade.Framework.csproj", "{6879932A-09D8-4F6F-99B3-52DF1349EE4C}"
 EndProject
@@ -23,7 +23,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Tardigrade.Framework.AuditN
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Tardigrade.Framework.ZXingNet", "Tardigrade.Framework.ZXingNet\Tardigrade.Framework.ZXingNet.csproj", "{947CB431-E5DE-40ED-AE94-153BF999AF45}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Tardigrade.Framework.MailKit", "Tardigrade.Framework.MailKit\Tardigrade.Framework.MailKit.csproj", "{BB7B60D8-E9BB-4743-B934-A0A102C1E11D}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Tardigrade.Framework.MailKit", "Tardigrade.Framework.MailKit\Tardigrade.Framework.MailKit.csproj", "{BB7B60D8-E9BB-4743-B934-A0A102C1E11D}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Tardigrade.Framework.AuditNET.AzureStorageQueue", "Tardigrade.Framework.AuditNET.AzureStorageQueue\Tardigrade.Framework.AuditNET.AzureStorageQueue.csproj", "{21019F4A-D35B-4007-A989-056608D3E9B8}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -75,6 +77,10 @@ Global
 		{BB7B60D8-E9BB-4743-B934-A0A102C1E11D}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{BB7B60D8-E9BB-4743-B934-A0A102C1E11D}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{BB7B60D8-E9BB-4743-B934-A0A102C1E11D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{21019F4A-D35B-4007-A989-056608D3E9B8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{21019F4A-D35B-4007-A989-056608D3E9B8}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{21019F4A-D35B-4007-A989-056608D3E9B8}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{21019F4A-D35B-4007-A989-056608D3E9B8}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
